API structure and documentation
This commit represents a (hopefully) quite final version of Troy's pilot API.
While that does differ from BigJob's pilot API, this is what Troy will implement
in the first iteration.

While documentation of the API should be relatively complete, the implementation
is yet non-functional -- simply because the CPI classes (Interface classes) are
out of sync with the API evolution (again).

Note that this API also includes a new implementation of the Attribute
interface, which is close to the GFD.90 attribute package, and which supports
monitoring via callbacks.