fix instance data, attribute interface
Instance data are now simply attributes on the API instances, where pre-defined
(needs documentation), or are entries in the 'idata' property dict.

The attribute interface now supports the full SAGA API (modulo some small
inspection methods), and also the python dictionary syntax.  Documentation is
complete, and a simple test was added.

A