And behold, as this is Troy! So goeth forth and enjoyeth and codeth the adaptorias.
I declare this to be the final iteration of the Troy.v1 API and structure
(Version 2 is a different matter).  No matter if it is perfect or not, this is
what we will code against for the next couple if month, until we gather more
experience and have a clear and motivated path forward.

Note that updated package documentation / intro.